Add files via upload
When program is launched with ./a.out it will run in interactive mode, batch mode, or return an error and exit. The program will parse through the command line in interactive mode and keep track of the number of commands and dynamically allocate an array to store commands.

I'm having an issue when dynamically allocating enough space for the string to hold each command though.